Semua admin/user

+++ routes/web.php
@@ -11,6 +11,8 @@ use App\Http\Controllers\Admin\Laporan\LapkasmasjidController;
 use App\Http\Controllers\Admin\Admin_Users\AdminusersController;
 use App\Http\Controllers\Admin\Kas_Masjid\PengeluaranController;
 use App\Http\Controllers\Admin\Admin_Profile_Setting\ProfilesettingController;
+use App\Http\Controllers\Admin\Kas_Masjid\SaldoweekController;
+use App\Http\Controllers\Admin\Photo_Masjid\PhotoController;
 
 /*
 |--------------------------------------------------------------------------
@@ -23,10 +25,13 @@ use App\Http\Controllers\Admin\Admin_Profile_Setting\ProfilesettingController;
 |
 */
 
+
 Route::get('/', [KasController::class,'index']);
 // Route::get('/', function () {
 //     return view('User.PartialsUser.home');
 // });
+
+
 Route::get('/about', function () {
     return view('User.About Us.about');
 });
@@ -56,6 +61,8 @@ Route::group(['prefix' => 'admin', 'middleware' => ['auth']], function () {
     // Route::get('/admin-users/edit/{id}', [AdminusersController::class,'update']);
     // Route::get('/admin-users/hapus/{id}', [AdminusersController::class,'destroy']);
     Route::resource('/users', AdminusersController::class)->middleware('admin:admin');
+    
+    Route::resource('/photos', PhotoController::class)->middleware('admin:admin');
     // Profile Setting
     // Route::resource('/profile-setting', ProfilesettingController::class);
     Route::get('/profile-setting', [ProfilesettingController::class,'index']);
@@ -65,21 +72,28 @@ Route::group(['prefix' => 'admin', 'middleware' => ['auth']], function () {
 
     Route::get('/kas-masjid/pemasukan', [PemasukanController::class, 'index'])->middleware('admin:admin|bendahara');
     Route::get('/kas-masjid/pengeluaran', [PengeluaranController::class, 'index'])->middleware('admin:admin|bendahara');
+    Route::get('/kas-masjid/saldo', [SaldoweekController::class,'index'])->middleware('admin:admin|bendahara');
     Route::get('/kas-masjid/rekap', [RekapController::class, 'index']);
-
     //AJAX JSON
     //Pemasukan
-    Route::get('/data-pemasukan', [PemasukanController::class, 'datapemasukan']);
-    Route::post('/data-pemasukan/add', [PemasukanController::class, 'store']);
-    Route::get('/data-pemasukan/edit/{id}', [PemasukanController::class, 'edit']);
-    Route::put('/data-pemasukan/update/{id}', [PemasukanController::class, 'update']);
-    Route::delete('/data-pemasukan/delete{id}', [PemasukanController::class, 'destroy']);
+    Route::get('/data-pemasukan', [PemasukanController::class, 'datapemasukan'])->middleware('admin:admin|bendahara');
+    Route::post('/data-pemasukan/add', [PemasukanController::class, 'store'])->middleware('admin:admin|bendahara');
+    Route::get('/data-pemasukan/edit/{id}', [PemasukanController::class, 'edit'])->middleware('admin:admin|bendahara');
+    Route::put('/data-pemasukan/update/{id}', [PemasukanController::class, 'update'])->middleware('admin:admin|bendahara');
+    Route::delete('/data-pemasukan/delete/{id}', [PemasukanController::class, 'destroy'])->middleware('admin:admin|bendahara');
     //Pengeluaran
-    Route::get('/data-pengeluaran', [PengeluaranController::class, 'datapengeluaran']);
-    Route::post('/data-pengeluaran/add', [PengeluaranController::class, 'store']);
-    Route::get('/data-pengeluaran/edit/{id}', [PengeluaranController::class, 'edit']);
-    Route::put('/data-pengeluaran/update/{id}', [PengeluaranController::class, 'update']);
-    Route::delete('/data-pengeluaran/delete/{id}', [PengeluaranController::class, 'destroy']);
+    Route::get('/data-pengeluaran', [PengeluaranController::class, 'datapengeluaran'])->middleware('admin:admin|bendahara');
+    Route::post('/data-pengeluaran/add', [PengeluaranController::class, 'store'])->middleware('admin:admin|bendahara');
+    Route::get('/data-pengeluaran/edit/{id}', [PengeluaranController::class, 'edit'])->middleware('admin:admin|bendahara');
+    Route::put('/data-pengeluaran/update/{id}', [PengeluaranController::class, 'update'])->middleware('admin:admin|bendahara');
+    Route::delete('/data-pengeluaran/delete/{id}', [PengeluaranController::class, 'destroy'])->middleware('admin:admin|bendahara');
+    //Saldo / minggu
+    Route::get('/data-saldo', [SaldoweekController::class,'datasaldo'])->middleware('admin:admin|bendahara');
+    Route::post('/data-saldo/add', [SaldoweekController::class,'store'])->middleware('admin:admin|bendahara');
+    Route::get('/data-saldo/edit/{id}', [SaldoweekController::class,'edit'])->middleware('admin:admin|bendahara');
+    Route::put('/data-saldo/update/{id}', [SaldoweekController::class,'update'])->middleware('admin:admin|bendahara');
+    Route::delete('/data-saldo/delete/{id}', [SaldoweekController::class, 'destroy'])->middleware('admin:admin|bendahara');
+    
     // Rekap kas-masjid
     Route::get('/data-rekap', [RekapController::class, 'datarekap']);
